requirements.txt
cloudscraper==1.2.71
python-telegram-bot==20.7
python-dotenv==1.0.0
flask==3.0.0


